directly upload images + pdfs
